Add migrations for config cleaning & update countries.json path
- Added migration 0015 to clean invalid country codes from existing data
- Added migration 0016 to alter the allowed_countries field with validation
- Refactored code and change code path as mentioned
