+++ buffalogs/impossible_travel/views/utils.py
@@ -24,17 +24,3 @@ def aggregate_alerts_interval(start_date, end_date, interval, date_fmt):
         aggregated_data[current_date.strftime(date_fmt)] = count
         current_date = next_date
     return aggregated_data
-
-
-def get_country_validation_sets():
-    """
-    Loads country data from countries.json and returns a tuple:
-    (VALID_COUNTRY_NAMES, VALID_COUNTRY_CODES)
-    """
-    try:
-        data = load_data("countries")
-        valid_country_names = set(data.keys())
-        valid_country_codes = set(data.values())
-        return valid_country_names, valid_country_codes
-    except FileNotFoundError:
-        return set(), set()
